:truck: rename : user 패키지를 member 패키지로 이름 변경
