+++ src/main/java/com/salayo/locallifebackend/domain/member/controller/MemberController.java
@@ -1,0 +1,5 @@
+package com.salayo.locallifebackend.domain.member.controller;
+
+public class MemberController {
+
+}
