+++ src/main/java/com/salayo/locallifebackend/domain/member/entity/Member.java
@@ -1,7 +1,7 @@
-package com.salayo.locallifebackend.domain.user.entity;
+package com.salayo.locallifebackend.domain.member.entity;
 
-import com.salayo.locallifebackend.domain.user.enums.Gender;
-import com.salayo.locallifebackend.domain.user.enums.MemberRole;
+import com.salayo.locallifebackend.domain.member.enums.Gender;
+import com.salayo.locallifebackend.domain.member.enums.MemberRole;
 import com.salayo.locallifebackend.global.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
