+++ src/main/java/com/salayo/locallifebackend/domain/member/enums/Gender.java
@@ -1,0 +1,7 @@
+package com.salayo.locallifebackend.domain.member.enums;
+
+public enum Gender {
+
+    MALE,
+    FEMALE
+}
