+++ src/main/java/com/salayo/locallifebackend/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.salayo.locallifebackend.domain.user.repository;
+package com.salayo.locallifebackend.domain.member.repository;
 
-import com.salayo.locallifebackend.domain.user.entity.Member;
+import com.salayo.locallifebackend.domain.member.entity.Member;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
