+++ src/main/java/com/salayo/locallifebackend/domain/member/service/MemberService.java
@@ -1,0 +1,5 @@
+package com.salayo.locallifebackend.domain.member.service;
+
+public class MemberService {
+
+}
