--- src/main/java/com/salayo/locallifebackend/domain/user/controller/MemberController.java
@@ -1,5 +1,0 @@
-package com.salayo.locallifebackend.domain.user.controller;
-
-public class MemberController {
-
-}
