--- src/main/java/com/salayo/locallifebackend/domain/user/enums/Gender.java
@@ -1,7 +1,0 @@
-package com.salayo.locallifebackend.domain.user.enums;
-
-public enum Gender {
-
-    MALE,
-    FEMALE
-}
