--- src/main/java/com/salayo/locallifebackend/domain/user/service/MemberService.java
@@ -1,5 +1,0 @@
-package com.salayo.locallifebackend.domain.user.service;
-
-public class MemberService {
-
-}
