+++ src/main/java/com/salayo/locallifebackend/global/security/filter/JwtAuthFilter.java
@@ -3,8 +3,8 @@ package com.salayo.locallifebackend.global.security.filter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.salayo.locallifebackend.domain.user.entity.Member;
-import com.salayo.locallifebackend.domain.user.repository.MemberRepository;
+import com.salayo.locallifebackend.domain.member.entity.Member;
+import com.salayo.locallifebackend.domain.member.repository.MemberRepository;
 import com.salayo.locallifebackend.global.error.ErrorCode;
 import com.salayo.locallifebackend.global.error.response.ErrorResponse;
 import com.salayo.locallifebackend.global.security.MemberDetails;
